fix: cross-platform CI fixes and &mut reference codegen
Windows/macOS CI:
- CMakeLists.txt: wrap -O1 in generator expression for MSVC /O1
- fluxqt CMakeLists.txt: platform-aware .dylib/.so, CMAKE_DL_LIBS, Threads
- tooling.cpp: .dylib extension check, platform-guarded linker flags
- fileio_runtime.cpp: version-gated CURLOPT_PROTOCOLS_STR for CURL >= 7.85

-Werror warnings:
- surrogate.cpp: int j → Eigen::Index j (signed/unsigned)
- control.cpp: static_cast<int>(.size()) for signed/unsigned
- analysis_runtime.cpp: static_cast<size_t>(std::max(...))
- model_quality_engine.cpp: C-casts → static_cast
- ast.h: remove std::move on double (pessimizing-move)

Runtime fixes:
- registration.cpp: wrap in namespace Flux {} (undefined symbol)
- codegen.cpp: fix &mut reference codegen — return existing alloca
  instead of copy, handle *ptr on LHS of assignment, unwrap Ref
  type in MemberExprAST and AssignExprAST for struct fields

382/423 integration tests pass, 0 failures

+++ src/runtime/analysis/analysis_runtime.cpp
@@ -256,7 +256,7 @@ extern "C" double flux_bode_analyze(double output_dbl, double freqStart, double 
     // Print magnitude table
     printf("  Freq (Hz)  |  Mag (dB)  | Phase (deg)\n");
     printf("  -----------+-----------+-------------\n");
-    for (size_t i = 0; i < freqs.size(); i += std::max(1, numPoints / 20)) {
+    for (size_t i = 0; i < freqs.size(); i += static_cast<size_t>(std::max(1, numPoints / 20))) {
         printf("  %9.2f  |  %+8.2f  |  %+8.2f\n", freqs[i], magDB[i], phaseDeg[i]);
     }
 
@@ -270,7 +270,7 @@ extern "C" double flux_bode_analyze(double output_dbl, double freqStart, double 
         double level = magMax - (row + 0.5) * magRange / 10.0;
         printf("  %+6.1f |", level);
         for (size_t i = 0; i < freqs.size(); i++) {
-            if (i % std::max(1, numPoints / 60) != 0) continue;
+            if (i % static_cast<size_t>(std::max(1, numPoints / 60)) != 0) continue;
             if (magDB[i] >= level)
                 printf("*");
             else
@@ -279,7 +279,7 @@ extern "C" double flux_bode_analyze(double output_dbl, double freqStart, double 
         printf("\n");
     }
     printf("  --------+");
-    for (size_t i = 0; i < freqs.size(); i += std::max(1, numPoints / 60))
+    for (size_t i = 0; i < freqs.size(); i += static_cast<size_t>(std::max(1, numPoints / 60)))
         printf("-");
     printf("\n");
 
@@ -293,7 +293,7 @@ extern "C" double flux_bode_analyze(double output_dbl, double freqStart, double 
         double level = phaseMax - (row + 0.5) * phaseRange / 10.0;
         printf("  %+6.1f |", level);
         for (size_t i = 0; i < freqs.size(); i++) {
-            if (i % std::max(1, numPoints / 60) != 0) continue;
+            if (i % static_cast<size_t>(std::max(1, numPoints / 60)) != 0) continue;
             if (phaseDeg[i] >= level)
                 printf("*");
             else
@@ -302,7 +302,7 @@ extern "C" double flux_bode_analyze(double output_dbl, double freqStart, double 
         printf("\n");
     }
     printf("  --------+");
-    for (size_t i = 0; i < freqs.size(); i += std::max(1, numPoints / 60))
+    for (size_t i = 0; i < freqs.size(); i += static_cast<size_t>(std::max(1, numPoints / 60)))
         printf("-");
     printf("\n\n");
